requirements.txt
fastapi~=0.115.7
uvicorn
sqlalchemy~=2.0.36
psycopg2-binary
pydantic~=2.10.6
python-dotenv